Move EGW1 to higher-level file.
